refactor(editor): split monolithic editor into faces/routes pages
- Extract shared components: RouteCard, ProgressRing, FullscreenTopoEditor, CragSelector
- Extract shared hook: useCragRoutes (crag + route loading logic)
- Extract shared utils: editor-utils (constants + preloadImage)
- Create /editor/faces page for face photo management
- Create /editor/routes page for topo line annotation
- Replace 2000+ line editor with lightweight hub page
- Move /demo/editor → /editor (no longer demo)
- Update all path references in source and docs
