+++ PROJECT_INDEX.md
@@ -20,7 +20,7 @@ src/
 │   │   ├── route/                # 线路列表页
 │   │   ├── profile/              # 设置页面
 │   │   ├── offline/              # 离线模式页面
-│   │   └── demo/editor/          # Topo 编辑器 (隐藏入口)
+│   │   └── editor/               # Topo 编辑器 (隐藏入口)
 │   ├── api/                      # API Routes (12 个端点)
 │   │   ├── beta/                 # Beta 视频 CRUD
 │   │   ├── weather/              # 天气数据 (高德API)
@@ -78,7 +78,7 @@ src/
 | 线路列表 | `src/app/[locale]/route/page.tsx` | 全部线路搜索 |
 | 设置 | `src/app/[locale]/profile/page.tsx` | 用户设置 |
 | 离线 | `src/app/[locale]/offline/page.tsx` | 已下载岩场 |
-| **Topo 编辑器** | `src/app/[locale]/demo/editor/page.tsx` | 线路标注 (隐藏入口) |
+| **Topo 编辑器** | `src/app/[locale]/editor/page.tsx` | 线路标注 (隐藏入口) |
 
 ### 🔐 隐藏入口
 **Topo 编辑器**：在任意页面 **2秒内连续点击 Tabbar "线路"按钮 6 次** 即可进入
