+++ src/components/editor/progress-ring.tsx
@@ -1,0 +1,51 @@
+/**
+ * 进度环组件
+ */
+export function ProgressRing({
+  progress,
+  size = 48,
+  strokeWidth = 4,
+}: {
+  progress: number
+  size?: number
+  strokeWidth?: number
+}) {
+  const radius = (size - strokeWidth) / 2
+  const circumference = radius * 2 * Math.PI
+  const offset = circumference - (progress / 100) * circumference
+
+  return (
+    <div className="relative" style={{ width: size, height: size }}>
+      <svg width={size} height={size} className="-rotate-90">
+        {/* 背景圆 */}
+        <circle
+          cx={size / 2}
+          cy={size / 2}
+          r={radius}
+          fill="none"
+          stroke="var(--theme-outline-variant)"
+          strokeWidth={strokeWidth}
+        />
+        {/* 进度圆 */}
+        <circle
+          cx={size / 2}
+          cy={size / 2}
+          r={radius}
+          fill="none"
+          stroke="var(--theme-primary)"
+          strokeWidth={strokeWidth}
+          strokeLinecap="round"
+          strokeDasharray={circumference}
+          strokeDashoffset={offset}
+          className="transition-all duration-500 ease-out"
+        />
+      </svg>
+      <div
+        className="absolute inset-0 flex items-center justify-center text-xs font-bold"
+        style={{ color: 'var(--theme-primary)' }}
+      >
+        {Math.round(progress)}%
+      </div>
+    </div>
+  )
+}
